refactor(app): split 2543-line controller.go into per-concern files
controller.go violated the 500-line limit (docs/go-codebase-checklist.md:283) at
5x over, with applyLocked an ~860-line function. Pure code reorganization, zero
logic change — moved functions into same-package files following the existing
detail.go/list.go/navigate.go pattern:

  controller.go 2543 -> 334  (struct, New, Apply, thin applyLocked dispatcher)
  menu.go        433  menu state/body/accessors
  footer.go      189  list/text/detail footer hint builders
  snapshot.go    175  Snapshot + body builders (help/identity)
  intents.go     200  ApplyIntents/applyIntents
  handle.go      214  Handle + result-lane handlers
  navigate.go    338  (existing) + applyNavResult/dispatchRelatedNavigate/applyRelatedNavResult
  actions.go     887  32 handleActionX methods (applyLocked switch broken into methods)

Verified: ready-to-push green, goldens byte-identical, lint 0, govulncheck clean.

actions.go (887), list.go (1619), detail.go (978) still exceed 500 — split next.
