+++ internal/app/menu.go
@@ -1,0 +1,433 @@
+package app
+
+import (
+	"maps"
+	"strings"
+
+	"github.com/k2m30/a9s/v3/internal/resource"
+	"github.com/k2m30/a9s/v3/internal/runtime"
+)
+
+// topMenuState returns the MenuState of the top-of-stack screen if it is
+// ScreenMenu, or nil otherwise.
+func (c *Controller) topMenuState() *MenuState {
+	if len(c.stack) == 0 {
+		return nil
+	}
+	top := c.stack[len(c.stack)-1]
+	if top.ID != runtime.ScreenMenu {
+		return nil
+	}
+	return c.stack[len(c.stack)-1].State.Menu
+}
+
+// rootMenuState returns the MenuState of the root (bottom) screen if it is
+// ScreenMenu. Menu intents always target the root menu regardless of which
+// screen is currently on top.
+func (c *Controller) rootMenuState() *MenuState {
+	if len(c.stack) == 0 {
+		return nil
+	}
+	if c.stack[0].ID != runtime.ScreenMenu {
+		return nil
+	}
+	return c.stack[0].State.Menu
+}
+
+// hasResourceScreen reports whether the stack contains at least one screen that
+// can initiate a reveal (ScreenResourceList, ScreenChildList, or ScreenDetail).
+// Used to guard spurious ValueRevealed events delivered when the only screen
+// visible is the menu (e.g. late delivery after a profile/region switch).
+// Caller must hold c.mu.
+func (c *Controller) hasResourceScreen() bool {
+	for _, s := range c.stack {
+		if s.ID == runtime.ScreenResourceList ||
+			s.ID == runtime.ScreenChildList ||
+			s.ID == runtime.ScreenDetail {
+			return true
+		}
+	}
+	return false
+}
+
+// menuVisibleItems returns the resource types visible under the current
+// MenuState filter + attention settings, mirroring mainmenu.go applyFilter.
+//
+// PR-C 1b: converge with mainmenu.go applyFilter + isVisibleUnderIssueFilter.
+func menuVisibleItems(ms *MenuState, all []resource.ResourceTypeDef) []resource.ResourceTypeDef {
+	var result []resource.ResourceTypeDef
+	if len(ms.Filter) < 2 {
+		result = all
+	} else {
+		q := strings.ToLower(ms.Filter)
+		result = make([]resource.ResourceTypeDef, 0, len(all))
+		for _, item := range all {
+			if strings.Contains(strings.ToLower(item.Name), q) ||
+				strings.Contains(strings.ToLower(item.ShortName), q) {
+				result = append(result, item)
+			}
+		}
+	}
+
+	if ms.AttentionOnly {
+		filtered := make([]resource.ResourceTypeDef, 0, len(result))
+		for _, item := range result {
+			if menuIsVisibleUnderIssueFilter(ms, item, menuActiveKey(ms, item)) {
+				filtered = append(filtered, item)
+			}
+		}
+		result = filtered
+	}
+	return result
+}
+
+// menuIsVisibleUnderIssueFilter mirrors mainmenu.go isVisibleUnderIssueFilter.
+//
+// item is the catalog entry; activeKey is the key under which intent data is
+// stored for this item (from menuActiveKey — may be an alias like "rds" for
+// the "dbi" type).
+//
+// PR-C 1b: converge with mainmenu.go isVisibleUnderIssueFilter.
+func menuIsVisibleUnderIssueFilter(ms *MenuState, item resource.ResourceTypeDef, activeKey string) bool {
+	known := ms.IssueKnown != nil && ms.IssueKnown[activeKey]
+	// ExcludeFromIssueBadge types are never probed — hide them in attention mode,
+	// even at cold-start, UNLESS issue data was explicitly recorded for them (a
+	// real detected issue beats the exclusion). In production these types are
+	// never probed, so this is equivalent to an absolute exclusion; the
+	// conditional only matters for tests that inject issues directly.
+	if item.ExcludeFromIssueBadge && !known {
+		return false
+	}
+	// Unknown non-excluded type: visible only during true cold-start (no type
+	// probed anywhere); once any probe lands, unknown types hide.
+	if !known {
+		return len(ms.IssueKnown) == 0
+	}
+	if ms.IssueCounts != nil && ms.IssueCounts[activeKey] > 0 {
+		return true
+	}
+	return ms.IssueTruncated != nil && ms.IssueTruncated[activeKey]
+}
+
+// menuSkipUnavailable advances the cursor past confirmed-empty resource types,
+// mirroring mainmenu.go skipUnavailable.
+//
+// PR-C 1b: converge with mainmenu.go skipUnavailable.
+func menuSkipUnavailable(ms *MenuState, visible []resource.ResourceTypeDef, direction int) {
+	if ms.Availability == nil || len(visible) == 0 {
+		return
+	}
+	total := len(visible)
+	start := ms.Cursor
+
+	cur := start
+	for cur >= 0 && cur < total {
+		item := visible[cur]
+		key := menuActiveKey(ms, item)
+		isTruncated := ms.Truncated != nil && ms.Truncated[key]
+		if count, known := ms.Availability[key]; !known || count > 0 || isTruncated {
+			ms.Cursor = cur
+			return
+		}
+		cur += direction
+	}
+
+	cur = start - direction
+	for cur >= 0 && cur < total {
+		item := visible[cur]
+		key := menuActiveKey(ms, item)
+		isTruncated := ms.Truncated != nil && ms.Truncated[key]
+		if count, known := ms.Availability[key]; !known || count > 0 || isTruncated {
+			ms.Cursor = cur
+			return
+		}
+		cur -= direction
+	}
+}
+
+// menuActiveKey returns the key under which intent data for the given
+// ResourceTypeDef is stored in a MenuState map. Intents are stored under
+// whatever key the runtime emits (e.g. "rds" for the "dbi" type); this
+// function resolves that key by checking the item's ShortName and all its
+// Aliases in order, returning the first one present in any of the three
+// intent maps. Falls back to item.ShortName when nothing is found.
+//
+// This allows buildMenuBody to expose the same key as the intent used —
+// so MenuEntry.ShortName matches what the runtime and tests expect.
+func menuActiveKey(ms *MenuState, item resource.ResourceTypeDef) string {
+	candidates := make([]string, 0, 1+len(item.Aliases))
+	candidates = append(candidates, item.ShortName)
+	candidates = append(candidates, item.Aliases...)
+	for _, c := range candidates {
+		if ms.Availability != nil {
+			if _, ok := ms.Availability[c]; ok {
+				return c
+			}
+		}
+		if ms.IssueKnown != nil {
+			if ms.IssueKnown[c] {
+				return c
+			}
+		}
+	}
+	return item.ShortName
+}
+
+// buildMenuBody constructs a MenuBody from MenuState + the resource catalog.
+// Applies the same filter + attention + skip-unavailable + badge logic as
+// mainmenu.go View(), but produces renderer-agnostic data instead of styled
+// strings.
+//
+// PR-C 1b: converge with mainmenu.go View() + FrameTitle().
+func buildMenuBody(ms *MenuState) *MenuBody {
+	all := resource.AllResourceTypes()
+	visible := menuVisibleItems(ms, all)
+
+	cursor := ms.Cursor
+	if cursor >= len(visible) && len(visible) > 0 {
+		cursor = len(visible) - 1
+	}
+
+	entries := make([]MenuEntry, 0, len(visible))
+	for _, item := range visible {
+		// Resolve the key under which intent data was stored for this type.
+		// Intents may use an alias ("rds") rather than the canonical ShortName
+		// ("dbi"); menuActiveKey finds whichever key has data, falling back to
+		// item.ShortName when no intent has been received yet.
+		activeKey := menuActiveKey(ms, item)
+
+		alias := ":" + item.ShortName
+		if len(item.Aliases) > 0 {
+			alias = ":" + item.Aliases[0]
+		}
+
+		avail, availKnown := 0, false
+		if ms.Availability != nil {
+			avail, availKnown = ms.Availability[activeKey]
+		}
+		availTruncated := ms.Truncated != nil && ms.Truncated[activeKey]
+
+		badge := IssueBadge{}
+		if ms.IssueKnown != nil && ms.IssueKnown[activeKey] {
+			cnt := 0
+			if ms.IssueCounts != nil {
+				cnt = ms.IssueCounts[activeKey]
+			}
+			trunc := ms.IssueTruncated != nil && ms.IssueTruncated[activeKey]
+			badge = IssueBadge{Count: cnt, Truncated: trunc}
+		}
+
+		entries = append(entries, MenuEntry{
+			ShortName:      activeKey,
+			Display:        item.Name,
+			Alias:          alias,
+			Category:       item.Category,
+			IssueBadge:     badge,
+			Availability:   avail,
+			AvailKnown:     availKnown,
+			AvailTruncated: availTruncated,
+		})
+	}
+
+	return &MenuBody{
+		Entries:       entries,
+		Selected:      cursor,
+		Filter:        ms.Filter,
+		AttentionOnly: ms.AttentionOnly,
+		Progress:      menuProgressIndicator(ms),
+	}
+}
+
+// menuFrameTitle mirrors mainmenu.go FrameTitle().
+//
+// PR-C 1b: converge with mainmenu.go FrameTitle().
+func menuFrameTitle(ms *MenuState) string {
+	all := resource.AllResourceTypes()
+	total := len(all)
+	visible := menuVisibleItems(ms, all)
+	filtered := len(visible)
+
+	var title string
+	switch {
+	case ms.Filter != "" || ms.AttentionOnly:
+		title = "resource-types(" + itoa(filtered) + "/" + itoa(total) + ")"
+	default:
+		title = "resource-types(" + itoa(total) + ")"
+	}
+	if ms.AttentionOnly {
+		title += " [!]"
+	}
+	if ms.EnrichTotal > 0 && ms.EnrichChecked < ms.EnrichTotal {
+		title += " [enriching " + itoa(ms.EnrichChecked) + "/" + itoa(ms.EnrichTotal) + "]"
+	}
+	return title
+}
+
+// menuProgressIndicator returns the scan/enrichment progress suffix only —
+// empty when no scan is active. This is what MenuBody.Progress carries;
+// menuFrameTitle() carries the full frame title string (base + suffix).
+func menuProgressIndicator(ms *MenuState) string {
+	if ms.EnrichTotal > 0 && ms.EnrichChecked < ms.EnrichTotal {
+		return "[enriching " + itoa(ms.EnrichChecked) + "/" + itoa(ms.EnrichTotal) + "]"
+	}
+	if ms.AvailTotal > 0 && ms.AvailChecked < ms.AvailTotal {
+		return "[checking " + itoa(ms.AvailChecked) + "/" + itoa(ms.AvailTotal) + "]"
+	}
+	return ""
+}
+
+// menuPageSize is the default cursor jump for PageUp/PageDown when the renderer
+// does not supply its viewport page size. The controller is renderer-neutral and
+// has no terminal height; 10 matches a typical visible window.
+const menuPageSize = 10
+
+// menuPageSizeFor returns the page size for a PageUp/PageDown action: the
+// renderer-supplied viewport page size (Action.N) when given, else the default.
+// The TUI passes max(height-1, 1) so page movement tracks the live viewport.
+func menuPageSizeFor(a Action) int {
+	if a.N > 0 {
+		return a.N
+	}
+	return menuPageSize
+}
+
+// itoa converts an int to its decimal string representation without importing
+// strconv (mirrors the views.itoa helper kept in the same conceptual layer).
+func itoa(n int) string {
+	if n == 0 {
+		return "0"
+	}
+	neg := n < 0
+	if neg {
+		n = -n
+	}
+	buf := [20]byte{}
+	pos := len(buf)
+	for n > 0 {
+		pos--
+		buf[pos] = byte('0' + n%10)
+		n /= 10
+	}
+	if neg {
+		pos--
+		buf[pos] = '-'
+	}
+	return string(buf[pos:])
+}
+
+// MenuFrameTitle returns the frame-border title for the main-menu screen,
+// delegating to menuFrameTitle with the root MenuState. Returns an empty
+// string when the root screen is not a menu.
+func (c *Controller) MenuFrameTitle() string {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	ms := c.rootMenuState()
+	if ms == nil {
+		return ""
+	}
+	return menuFrameTitle(ms)
+}
+
+// MenuSelected returns the ResourceTypeDef at the current cursor and a bool
+// that is true when navigation is permitted (i.e. the item is not confirmed
+// empty). Mirrors the Enter-key guard in ActionSelect.
+func (c *Controller) MenuSelected() (resource.ResourceTypeDef, bool) {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	ms := c.rootMenuState()
+	if ms == nil {
+		return resource.ResourceTypeDef{}, false
+	}
+	all := resource.AllResourceTypes()
+	visible := menuVisibleItems(ms, all)
+	if len(visible) == 0 {
+		return resource.ResourceTypeDef{}, false
+	}
+	// Background issue/availability intents can shrink the visible list while the
+	// stored cursor still points past the end. Snapshot clamps the displayed
+	// selection to the last visible row, so clamp here too — otherwise Enter on
+	// the highlighted last row would hit a stale guard and become a no-op.
+	cursor := ms.Cursor
+	if cursor >= len(visible) {
+		cursor = len(visible) - 1
+	}
+	selected := visible[cursor]
+	if ms.Availability != nil {
+		key := menuActiveKey(ms, selected)
+		isTruncated := ms.Truncated != nil && ms.Truncated[key]
+		if count, known := ms.Availability[key]; known && count == 0 && !isTruncated {
+			return selected, false
+		}
+	}
+	return selected, true
+}
+
+// GetMenuAvailability returns a copy of the root MenuState availability map.
+// Returns nil when no availability data has been recorded.
+func (c *Controller) GetMenuAvailability() map[string]int {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	ms := c.rootMenuState()
+	if ms == nil || ms.Availability == nil {
+		return nil
+	}
+	cp := make(map[string]int, len(ms.Availability))
+	maps.Copy(cp, ms.Availability)
+	return cp
+}
+
+// GetMenuTruncated returns a copy of the root MenuState truncated map.
+// Returns nil when no truncation data has been recorded.
+func (c *Controller) GetMenuTruncated() map[string]bool {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	ms := c.rootMenuState()
+	if ms == nil || ms.Truncated == nil {
+		return nil
+	}
+	cp := make(map[string]bool, len(ms.Truncated))
+	maps.Copy(cp, ms.Truncated)
+	return cp
+}
+
+// GetMenuIssueCounts returns a copy of the root MenuState issue-count map.
+// Returns nil when no issue data has been recorded.
+func (c *Controller) GetMenuIssueCounts() map[string]int {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	ms := c.rootMenuState()
+	if ms == nil || ms.IssueCounts == nil {
+		return nil
+	}
+	cp := make(map[string]int, len(ms.IssueCounts))
+	maps.Copy(cp, ms.IssueCounts)
+	return cp
+}
+
+// GetMenuIssueKnown returns a copy of the root MenuState issue-known map.
+// Returns nil when no issue data has been recorded.
+func (c *Controller) GetMenuIssueKnown() map[string]bool {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	ms := c.rootMenuState()
+	if ms == nil || ms.IssueKnown == nil {
+		return nil
+	}
+	cp := make(map[string]bool, len(ms.IssueKnown))
+	maps.Copy(cp, ms.IssueKnown)
+	return cp
+}
+
+// GetMenuIssueTruncated returns a copy of the root MenuState issue-truncated map.
+// Returns nil when no issue-truncation data has been recorded.
+func (c *Controller) GetMenuIssueTruncated() map[string]bool {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	ms := c.rootMenuState()
+	if ms == nil || ms.IssueTruncated == nil {
+		return nil
+	}
+	cp := make(map[string]bool, len(ms.IssueTruncated))
+	maps.Copy(cp, ms.IssueTruncated)
+	return cp
+}
