+++ internal/app/navigate.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/k2m30/a9s/v3/internal/fieldpath"
 	"github.com/k2m30/a9s/v3/internal/resource"
+	"github.com/k2m30/a9s/v3/internal/runtime"
 	"gopkg.in/yaml.v3"
 )
 
@@ -56,5 +57,282 @@ func resourceJSONLines(r resource.Resource) []string {
 	}
 	raw := strings.TrimRight(string(data), "\n")
 	return strings.Split(raw, "\n")
+}
+
+// applyNavResult converts a NavigateResult into PushScreen/ReplaceScreen/PopScreen
+// stack operations. Called by Apply after HandleNavigate returns.
+//
+// The adapter (not the runtime) decides which ScreenID to push for each kind;
+// this method encodes that mapping for the headless controller.
+//
+// Kinds that require selected-row or resource data (PushDetail, PushYAML,
+// PushJSON, PushResourceList/Cached, FetchReveal) are deferred to PR-C where
+// per-screen state is lifted into the controller.
+func (c *Controller) applyNavResult(res runtime.NavigateResult) {
+	switch res.Kind {
+	case runtime.NavigateKindPopAll:
+		// Pop back to the root menu — leave exactly one screen, never empty.
+		// (Popping via ApplyIntents would now stop at the len<=1 guard anyway;
+		// pop directly to keep the intent clear.)
+		if len(c.stack) > 1 {
+			c.stack = c.stack[:1]
+		}
+
+	case runtime.NavigateKindPushHelp:
+		c.applyIntents([]runtime.UIIntent{runtime.PushScreen{ID: runtime.ScreenHelp}})
+
+	case runtime.NavigateKindPushRegion:
+		c.applyIntents([]runtime.UIIntent{runtime.PushScreen{ID: runtime.ScreenRegion}})
+
+	case runtime.NavigateKindPushTheme:
+		c.applyIntents([]runtime.UIIntent{runtime.PushScreen{ID: runtime.ScreenTheme}})
+
+	case runtime.NavigateKindFetchProfiles:
+		// No stack change — the adapter starts the fetch task; when the result
+		// arrives (ProfilesLoaded), HandleProfilesLoaded pushes ScreenProfileSelector.
+
+	case runtime.NavigateKindFlash, runtime.NavigateKindNoop:
+		// No stack change — flash is surfaced via FlashIntent in the intent stream.
+
+	case runtime.NavigateKindPushResourceList, runtime.NavigateKindPushResourceListCached:
+		intent := runtime.PushScreen{
+			ID:      runtime.ScreenResourceList,
+			Context: runtime.ScreenContext{ResourceType: res.ResolvedType},
+		}
+		if res.ReplaceCurrent {
+			c.applyIntents([]runtime.UIIntent{runtime.ReplaceScreen{ID: intent.ID, Context: intent.Context}})
+		} else {
+			c.applyIntents([]runtime.UIIntent{intent})
+		}
+		c.ensureListState()
+		// For the cached path, populate rows immediately from the cache entry so
+		// headless/web callers see data without waiting for a fetch round-trip.
+		if res.Kind == runtime.NavigateKindPushResourceListCached && res.CachedEntry != nil {
+			top := &c.stack[len(c.stack)-1]
+			c.applyResourcesLoaded(top.State.List, res.ResolvedType, res.CachedEntry.Resources, res.CachedEntry.Pagination, false)
+		}
+
+	case runtime.NavigateKindPushDetail:
+		if res.Resource == nil {
+			return
+		}
+		intent := runtime.PushScreen{
+			ID: runtime.ScreenDetail,
+			Context: runtime.ScreenContext{
+				ResourceType: res.ResolvedType,
+				ResourceID:   res.Resource.ID,
+			},
+		}
+		c.applyIntents([]runtime.UIIntent{intent})
+		// Use lock-free variants — applyNavResult is always called while c.mu
+		// is already held by Apply or Handle.
+		c.ensureDetailState(*res.Resource, res.ResolvedType)
+		if res.DispatchRelated {
+			c.initDetailRelatedRows(res.ResolvedType)
+		}
+
+	case runtime.NavigateKindPushYAML:
+		if res.Resource == nil {
+			return
+		}
+		lines := resourceYAMLLines(*res.Resource)
+		intent := runtime.PushScreen{
+			ID: runtime.ScreenYAML,
+			Context: runtime.ScreenContext{
+				ResourceType: res.ResolvedType,
+				ResourceID:   res.Resource.ID,
+			},
+		}
+		c.applyIntents([]runtime.UIIntent{intent})
+		c.ensureTextState(lines)
+
+	case runtime.NavigateKindPushJSON:
+		if res.Resource == nil {
+			return
+		}
+		lines := resourceJSONLines(*res.Resource)
+		intent := runtime.PushScreen{
+			ID: runtime.ScreenJSON,
+			Context: runtime.ScreenContext{
+				ResourceType: res.ResolvedType,
+				ResourceID:   res.Resource.ID,
+			},
+		}
+		c.applyIntents([]runtime.UIIntent{intent})
+		c.ensureTextState(lines)
+
+	case runtime.NavigateKindFetchReveal:
+		// No stack push yet — the push happens when Handle receives
+		// messages.ValueRevealed and routes it to HandleValueRevealed.
+		// Tasks are returned by Apply's ActionReveal branch directly.
+	}
+}
+
+// dispatchRelatedNavigate calls HandleRelatedNavigate then applyRelatedNavResult
+// and merges the two task slices, preferring extraTasks when the same Key
+// appears in both (applyRelatedNavResult returns payload-bearing replacements
+// for tasks HandleRelatedNavigate emits without payloads, e.g. KindFetchFiltered).
+// All three related-nav callers — ActionSelect keyboard, ActionRelatedSelect
+// click, and ActionFieldSelect click — use this shared tail.
+func (c *Controller) dispatchRelatedNavigate(ev runtime.RelatedNavigateEvent) []runtime.TaskRequest {
+	navRes, tasks := c.core.HandleRelatedNavigate(ev)
+	extraTasks := c.applyRelatedNavResult(navRes)
+	if len(extraTasks) == 0 {
+		return tasks
+	}
+	extraKeys := make(map[runtime.TaskKey]struct{}, len(extraTasks))
+	for _, t := range extraTasks {
+		extraKeys[t.Key] = struct{}{}
+	}
+	merged := make([]runtime.TaskRequest, 0, len(tasks)+len(extraTasks))
+	for _, t := range tasks {
+		if _, replaced := extraKeys[t.Key]; !replaced {
+			merged = append(merged, t)
+		}
+	}
+	merged = append(merged, extraTasks...)
+	return merged
+}
+
+// applyRelatedNavResult converts a NavigationResult into stack operations and
+// returns any additional task requests the result spawns.
+//
+// NavigationResult carries NavigationKind plus TargetType, TargetID,
+// RelatedIDs, FetchFilter, FilterText — no ScreenID; the controller maps
+// kind → ScreenID.
+func (c *Controller) applyRelatedNavResult(res runtime.NavigationResult) []runtime.TaskRequest {
+	switch res.Kind {
+	case runtime.NavigationKindResourceList:
+		intent := runtime.PushScreen{
+			ID:      runtime.ScreenResourceList,
+			Context: runtime.ScreenContext{ResourceType: res.TargetType},
+		}
+		c.applyIntents([]runtime.UIIntent{intent})
+		c.ensureListState()
+
+	case runtime.NavigationKindFilteredList:
+		intent := runtime.PushScreen{
+			ID:      runtime.ScreenResourceList,
+			Context: runtime.ScreenContext{ResourceType: res.TargetType},
+		}
+		c.applyIntents([]runtime.UIIntent{intent})
+		c.ensureListState()
+		if ls := c.topListState(); ls != nil {
+			if res.FilterText != "" {
+				ls.Filter = res.FilterText
+			}
+			if len(res.FetchFilter) > 0 {
+				ls.FetchFilter = res.FetchFilter
+				// HandleRelatedNavigate returns a no-payload KindFetchFiltered task
+				// (tested as-is by the QA suite). Replace it here with a payload-
+				// bearing version so the executor can invoke the filtered fetcher.
+				return []runtime.TaskRequest{{
+					Key:     runtime.TaskKey{Kind: runtime.KindFetchFiltered, Scope: res.TargetType},
+					Cache:   runtime.CacheNone,
+					Payload: runtime.FetchFilteredPayload{Filter: res.FetchFilter},
+				}}
+			}
+			if len(res.RelatedIDs) > 0 {
+				// Multi-ID related nav with no server-side filter (e.g. an EC2
+				// instance → its several security groups): prefilter the list to
+				// just the related subset, mirroring the TUI related-list path.
+				// Without this the list shows every resource of the target type
+				// instead of the related ones. We're already under c.mu, so seed
+				// ls directly — PatchListRelatedIDSet would re-lock and deadlock.
+				set := make(map[string]struct{}, len(res.RelatedIDs))
+				for _, id := range res.RelatedIDs {
+					if id != "" {
+						set[id] = struct{}{}
+					}
+				}
+				ls.RelatedIDSet = set
+			}
+		}
+
+	case runtime.NavigationKindDetail:
+		// The target resource is already cached: NavigationKindDetail is only
+		// returned on a cache hit (ResolveRelatedNavigate), and HandleRelatedNavigate
+		// returns no fetch task for it ("the adapter serves these from cached
+		// state"). Seed the detail synchronously from the cache — this works for
+		// every type (most have no by-id fetcher, so a fetch would land on an empty
+		// detail). Mirrors ActionOpenDetail's related-panel handling.
+		id := res.TargetID
+		if id == "" && len(res.RelatedIDs) == 1 {
+			id = res.RelatedIDs[0]
+		}
+		cached, ok := c.core.RelatedCachedResource(res.TargetType, id)
+		if !ok {
+			// Defensive: cache unexpectedly missing. Land on a filtered list by id
+			// instead of an empty detail so navigation still resolves somewhere.
+			c.applyIntents([]runtime.UIIntent{runtime.PushScreen{
+				ID:      runtime.ScreenResourceList,
+				Context: runtime.ScreenContext{ResourceType: res.TargetType},
+			}})
+			c.ensureListState()
+			if ls := c.topListState(); ls != nil && id != "" {
+				ls.Filter = id
+			}
+			return nil
+		}
+		c.applyIntents([]runtime.UIIntent{runtime.PushScreen{
+			ID:      runtime.ScreenDetail,
+			Context: runtime.ScreenContext{ResourceType: res.TargetType, ResourceID: id},
+		}})
+		c.ensureDetailState(cached, res.TargetType)
+		ds := c.topDetailState()
+		if ds == nil || len(resource.GetRelated(res.TargetType)) == 0 {
+			return nil
+		}
+		c.initDetailRelatedRows(res.TargetType)
+		// Populate the related panel: replay the related cache if present, else
+		// dispatch a KindRelatedCheck task — same as ActionOpenDetail.
+		ck := runtime.RelatedCacheKey(res.TargetType, cached.ID)
+		if cachedRows, hit := c.core.RelatedCacheGet(ck); hit && len(cachedRows) > 0 {
+			for _, entry := range cachedRows {
+				errMsg := ""
+				if entry.Result.Err != nil {
+					errMsg = entry.Result.Err.Error()
+				}
+				mergeDetailRelatedRow(ds, entry.DefDisplayName, entry.Result.TargetType,
+					entry.Result.Count, false, errMsg, entry.Result.Approximate, entry.Result.ResourceIDs, entry.Result.FetchFilter)
+			}
+			return nil
+		}
+		return []runtime.TaskRequest{{
+			Key:     runtime.TaskKey{Kind: runtime.KindRelatedCheck, Scope: res.TargetType + "/" + cached.ID},
+			Cache:   runtime.CacheNone,
+			Payload: runtime.RelatedCheckPayload{ResourceType: res.TargetType, Resource: cached},
+		}}
+
+	case runtime.NavigationKindEnterChildView:
+		// Delegate to the same path used by ActionChildView but with the
+		// target type already resolved. Build a minimal EnterChildViewEvent.
+		ev := runtime.EnterChildViewEvent{
+			ChildType: res.TargetType,
+		}
+		intents, tasks := c.core.HandleEnterChildView(ev)
+		c.applyIntents(intents)
+		if len(c.stack) > 0 {
+			top := &c.stack[len(c.stack)-1]
+			if top.ID == runtime.ScreenChildList {
+				top.Ctx.ResourceType = res.TargetType
+				if top.State.List == nil {
+					top.State.List = &ListState{Loading: true}
+				}
+				if res.FetchFilter != nil {
+					top.State.List.ParentContext = res.FetchFilter
+				}
+			}
+		}
+		return tasks
+
+	case runtime.NavigationKindFlash:
+		// Flash is surfaced as a FlashIntent by HandleRelatedNavigate — no
+		// stack change needed here.
+
+	default:
+		// NavigationKindUnknown or future kinds: no-op.
+	}
+	return nil
 }
